refactor/structs: Struct extraction.
Move structs from core/ipc to core/core_structs.rs for clarity
