+++ safe_app/src/ffi/mutable_data/entries.rs
@@ -19,7 +19,7 @@ use ffi_utils::{
     catch_unwind_cb, vec_clone_from_raw_parts, FfiResult, OpaqueCtx, SafePtr, FFI_RESULT_OK,
 };
 use safe_core::ffi::ipc::resp::MDataEntry;
-use safe_core::ipc::resp::{
+use safe_core::core_structs::{
     MDataEntry as NativeMDataEntry, MDataKey as NativeMDataKey, MDataValue as NativeMDataValue,
 };
 use safe_core::CoreError;
@@ -239,7 +239,7 @@ mod tests {
         call_0, call_1, call_vec, send_via_user_data, sender_as_user_data,
     };
     use ffi_utils::vec_clone_from_raw_parts;
-    use safe_core::ipc::resp::{MDataEntry, MDataKey, MDataValue};
+    use safe_core::core_structs::{MDataEntry, MDataKey, MDataValue};
     use safe_core::utils;
     use safe_nd::{MDataAction, MDataPermissionSet, MDataSeqValue};
     use std::os::raw::c_void;
