+++ safe_app/src/ffi/mutable_data/mod.rs
@@ -36,7 +36,7 @@ use safe_core::ffi::ipc::resp::MDataKey;
 use safe_core::ffi::ipc::resp::MDataValue;
 use safe_core::ffi::MDataInfo;
 use safe_core::ipc::req::{permission_set_clone_from_repr_c, permission_set_into_repr_c};
-use safe_core::ipc::resp::{MDataKey as NativeMDataKey, MDataValue as NativeMDataValue};
+use safe_core::core_structs::{MDataKey as NativeMDataKey, MDataValue as NativeMDataValue};
 use safe_core::{Client, FutureExt, MDataInfo as NativeMDataInfo};
 use safe_nd::{
     EntryError, Error as SndError, MDataPermissionSet, MDataSeqEntryAction, MDataSeqValue,
