+++ safe_authenticator/src/apps.rs
@@ -20,8 +20,8 @@ use ffi_utils::{vec_into_raw_parts, ReprC};
 use futures::future::Future;
 use safe_core::client::Client;
 use safe_core::ipc::req::{containers_from_repr_c, containers_into_vec, ContainerPermissions};
-use safe_core::ipc::resp::{AccessContainerEntry, AppAccess};
-use safe_core::ipc::{access_container_enc_key, AppExchangeInfo, IpcError};
+use safe_core::core_structs::{AccessContainerEntry, AppAccess, access_container_enc_key};
+use safe_core::ipc::{AppExchangeInfo, IpcError};
 use safe_core::utils::symmetric_decrypt;
 use safe_core::FutureExt;
 use safe_nd::{AppPermissions, MDataAddress, XorName};
