+++ safe_authenticator/src/ffi/apps.rs
@@ -20,7 +20,7 @@ use safe_core::ffi::arrays::XorNameArray;
 use safe_core::ffi::ipc::req::{AppExchangeInfo, ContainerPermissions};
 use safe_core::ffi::ipc::resp::AppAccess;
 use safe_core::ipc::req::AppExchangeInfo as NativeAppExchangeInfo;
-use safe_core::ipc::resp::AppAccess as NativeAppAccess;
+use safe_core::core_structs::AppAccess as NativeAppAccess;
 use safe_core::FutureExt;
 use safe_nd::XorName;
 use std::os::raw::{c_char, c_void};
