+++ safe_authenticator/src/ipc.rs
@@ -19,7 +19,8 @@ use futures::future::{self, Either};
 use futures::Future;
 use safe_core::ffi::ipc::resp::MetadataResponse as FfiUserMetadata;
 use safe_core::ipc::req::{IpcReq, ShareMDataReq};
-use safe_core::ipc::resp::{IpcResp, UserMetadata, METADATA_KEY};
+use safe_core::ipc::resp::{IpcResp};
+use safe_core::core_structs::{UserMetadata, METADATA_KEY};
 use safe_core::ipc::{self, IpcError, IpcMsg};
 use safe_core::{Client, CoreError, FutureExt};
 use safe_nd::{Error as SndError, XorName};
