+++ safe_authenticator/src/tests/share_mdata.rs
@@ -16,7 +16,7 @@ use bincode::serialize;
 use ffi_utils::test_utils::{call_1, call_vec};
 use futures::Future;
 use safe_core::ipc::req::AppExchangeInfo;
-use safe_core::ipc::resp::{AppAccess, UserMetadata, METADATA_KEY};
+use safe_core::core_structs::{AppAccess, UserMetadata, METADATA_KEY};
 use safe_core::ipc::{self, AuthReq, IpcError, IpcMsg, IpcReq, IpcResp, ShareMData, ShareMDataReq};
 use safe_core::Client;
 use safe_nd::{MDataAction, MDataPermissionSet, MDataSeqValue, PublicKey, SeqMutableData};
